pyproject.toml
[project]
name = "rcv"
version = "0.1.0"
description = "Resume Control Versioning - A CLI tool for managing versioned resumes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
rcv = "rcv.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rcv"]
